Update the path to the aeolus-cli example configuration file
The path changed when we split out the cli from the gem library.
The image-factory module needs to make sure the rubygem-aeolus-cli
package is installed.  I left the rubygem-aeolus-image package installation
in tact in case other modules still need it.  Probably not great to leave
a transitive dependency in, but we can address it later.