BZ 749254 - running configure for multiple providers causes oauth issues.
The problem was aeolus-configure generated new uuids each time it is invoked.

Matt has created a patch for conductor to have the rpm generate and install
a oauth.json file which will be the canonical source for the keys.

This patch augments configure to read oauth.json and to produce custom facts
replacing the auto generated uuids in aeolus-configure.