BZ 753273 - rsyslog configuration causes messages log to grow rapidly.
https://bugzilla.redhat.com/show_bug.cgi?id=753273

There were a number of small things causing this:
* We were logging relp output to localhost
* We were logging it to messages instead of /var/log/aeolus.log
* There was a syntax error
* We were spinning up an input relp server on a port that we don't
  yet have an selinux policy for

I fixed the main issues and commented all the enabling bits out. The
user/admin really needs to put where to output, and they are unlikely
to actually use the input module, but it is harmless to leave as an
example.