BZ752975 - multiple executions of aeolus-cleanup/configure not working as expected
https://bugzilla.redhat.com/show_bug.cgi?id=752975

This occurs because providers that already exist are attempting to be
recreated.  We determine whether or not a provider exists by issuing a
GET on /conductor/providers.  This request issues a 302 response, but
we are not properly following that redirect due to looking for the
follow param in the wrong place.  This corrects that so the follow
param works as expected, and thus the resource is identified and not
wrongfully recreated.