BZ 758644 (part 1) - clean old cookies when doing a web request
If :store_cookies_at option is used and cookies file already exists,
this file is now truncated. Reason for this is that cookies file is
removed in logout action by default, but if aeolus-configure
is executed multiple times, logout action is sometimes not executed ->
cookie file is not deleted and is used in next aeolus-configure execution.

This patch truncates the cookies file anytime ':store_cookies_at' option is used
when calling web_request (it should be only on login request).