BZ 782210 - improved RHEV NFS export validation
https://bugzilla.redhat.com/show_bug.cgi?id=782210

For 782210, the RHEV-M REST API doesn't support queries on the storage
domain type.

So to do this properly, we now query the REST API to have it return the xml
containing a list of all storage domains in the data center. We then
perform a XPATH query to find a storage domain that has type equal to
'export' and that has the correct export path.

This patch also addresses the issues in:
https://bugzilla.redhat.com/show_bug.cgi?id=782545
https://bugzilla.redhat.com/show_bug.cgi?id=782543