BZ 785217 - check provider add success by inspecting the flash image alt text
https://bugzilla.redhat.com/show_bug.cgi?id=785217

Bug was produced when "Provider added." was changed to "Cloud Resource Provider added."

This text is internationalized so to simplify the check, we instead,
look at the alt text of the image that is included in the flash message.

The alt text has the value 'Notices' if there are no errors and
has the value 'Warnings' if there are errors.