BZ 794755 - Static assets don't set Cache-Control headers https://bugzilla.redhat.com/show_bug.cgi?id=794755
Use mod_expires to set expiration for css/js/png to one year.

From the BZ:

AssetTagHelper appends timestamps as a query parameter, so we can set a
Cache-Control directive allowing these assets to be cached forever -- if we
change a stylesheet, it will end up with a new timestamp in the URL and get
reloaded automatically.

Signed-off-by: Matt Wagner <matt.wagner@redhat.com>