BZ 802847 - cleanup hardware profiles and fix provider ordering effects
This revised patch removes the i386 hwp.

https://bugzilla.redhat.com/show_bug.cgi?id=802847

Reverted the hwp1 memory change. Setting the memory to 7500 is incompatible with vsphere. vsphere provider hwp memory range is 512-2048.

Setting memory back to 512m for the default profile. Once the conductor and deltacloud changes described in this patch is in.

https://fedorahosted.org/pipermail/aeolus-devel/2012-March/009611.html

We will be able to start using the small instances in ec2.

hwp1 has also been renamed to small-x86_64.