Fix broken string interpolation
(Part of https://www.aeolusproject.org/redmine/issues/3240)

String interpolation has changed in Puppet 2.7 regarding variables
with dashes in the name (although I can't find any documentation on
this).  Consider the following code:

  $foo = "FOO"
  $bar = "BAR"
  notice("$foo-$bar")

Under 2.6, this produces:

  notice: Scope(Class[main]): FOO-BAR

But under 2.7, this produces:

  notice: Scope(Class[main]): BAR

What is happening here is that the hyphen character is considered part
of the variable in 2.7; literally the variable is '$foo-', which is
not defined and evaluates to an empty string.  To be safe, always
enclose string-interpolated variables in curly braces as below:

  $foo = "FOO"
  $bar = "BAR"
  notice("${foo}-${bar}")

This will produce the expected results.