BZ 827562 - Upgrade from CloudForms 1.0 to CloudForms 1.0.1 requires additional data definition not provided by rails migration
We need a script to stop,start,restart aeolus services so that we
can stop them, insert data required by upgrade path, and restart.
The existing aeolus-services script appears deprecated and is used
nowhere else in the codebase for aeolus-configure or aeolus-conductor,
so we'll refactor it to be useful for this purpose.