BZ#806001 - Remove admin profile
The default admin user will be created as part of the conductor
manifest, and this will be indicated by creation of a file in
/var/lib/aeolus-configure. This will ensure the admin user does not
get created by configure after multiple configure invocations.  The
expectation is for administrative users to be managed via the web UI
(or API), not via configure.

Signed-off-by: John Eckersberg <jeckersb@redhat.com>