BZ#847399 - changes needed around admin users
https://bugzilla.redhat.com/show_bug.cgi?id=847399

This replaces destroy_temp_admin with destroy_temp_admins, which uses
the new dc:destroy_users_by_pattern rake task to destroy *all*
temporary users.

In addition, we now destroy temp admins twice during a given configure
run:
- Before attempting to create a site_admin user
- At the end, as before

Also, the site admin is created before the temp admin.  All of this
collectively ensures that the site admin is successfully created if
needed.  Before it was possible for the temp admin to be created first
(or for "stale" temp admins to be in the database) which would prevent
the site admin from being created.